Added challenge member search to progress dropdown
